Add variables to define alternative paths/filename
Variables are checked for existence beforehand, so you can set them
wherever you want in your vimrc.

This commit also adds a reference to the creator of the cfr Java
decompiler; this may not be necessary in the future, but I like to give
credit where credit is due.

The plugin also lcd's to the file rather than cd, since I thought that
might be annoying for the nested package navigation Java practices
usually encourage.